move gossiper to separate file
